Add activities_with_future_transactions to output
