+++ stats/analytics.py
@@ -1407,6 +1407,15 @@ class ActivityStats(CommonSharedElements):
             out[get_currency(self, pd)][planned_disbursement_year(pd)] += disbursement_value
         return out
 
+    @returns_numberdict
+    def activities_with_future_transactions(self):
+        future_transactions = 0
+        for transaction in self.element.findall('transaction'):
+            if transaction_date(transaction) > self.today:
+                future_transactions += 1
+        if future_transactions:
+            return {self.element.find('iati-identifier').text: future_transactions}
+
 
 ckan = json.load(open('helpers/ckan.json'))
 publisher_re = re.compile(r'(.*)\-[^\-]')
